requirements.txt
azure-functions
pymongo
feedparser
google-generativeai
azure-identity
azure-keyvault-secrets
structlog
python-dotenv
markdown
pytest
azure-storage-queue
langchain
langchain-community
langchain-google-genai
playwright
beautifulsoup4
lxml
requests